==> dataset/match_stats/03_consolidation.ipynb <==
import pandas as pd

# Joining the separate dataframes

Loading the original dataframes

df_matches = pd.read_csv("./initial_ds/match_stat_links.csv")
df_stats = pd.read_csv("./initial_ds/match_stats.csv")

The dataframe containing the matches and the index columns (date and team names):

df_matches

The dataframe containing match stats:

df_stats

Joining the two initial dataframes on their common column which is the URL of the match's stats (stats_link):

joined_df = pd.merge(left=df_matches, right=df_stats, on='stats_link', how='inner')

joined_df

# Cleaning up

joined_df.columns

joined_df = joined_df.drop(['season', 'link', 'stats_link'], axis='columns')

# Fixing the dates

joined_df['date']

joined_df['match_date']

joined_df['date'] = pd.to_datetime(joined_df['match_date']).dt.date

joined_df = joined_df.drop('match_date', axis='columns')

# Saving the results into file

joined_df.to_csv("./processed_ds/match_stats.csv", index=False)
